CLI: Implement basic CLI parser.
Introduces CLI parser and HTool Client.

** CLI PARSER **

This commit implements the very basic CLI parser. The --version and --help commands have
been added, both working, and the necessary code for actual operations like `file` or
`macho` are also present.

Adding new commands takes a few steps:

 1. Define the command and all it's options/arguments in an `struct option` array at
    the top of main.c. See existing for examples.
 2. Add the command name and handler function for the `commands` array.
 3. Define the command handler at the top of the main.c file. All handlers should be
    a. static
    b. have the `htool_return_t` return type,
    c. have the htool client, argc and argv as parameters.
 4. Write the handler function below the command options definitions (section is
    labelled). Must use the structure documented by getopt.

When writing command handlers, any options need to be added as flag definitions within
the htool.h header file. The command should be defined as a hex value with the most
significant bit set higher than the last command, and then a set of opt defines with
an appropriate hex value, e.g.

	#define HTOOL_CLIENT_CMDFLAG_TEST	0x40000000

	#define HTOOL_CLIENT_TEST_OPT_ALPHA	(1 << 1)
	#define HTOOL_CLIENT_TEST_OPT_BETA	(1 << 2)

The naming scheme for these defines may change in the future, but they're fine at the
moment.

** HTOOL CLIENT **

This commit implements part of the htool client API - used for handling information
throughout the program and avoiding too many function parameters.

Currently, the `htool_client_t` structure is implemented, with the filename, full arg
string, cmd/opts/arch flags, and the parsed binary pointer. More needs to be added to
this, including a function for quickly checking the flags - saving bitwise ops else-
where in the code.

---

Signed-off-by: Harry Moulton <me@h3adsh0tzz.com>