Loader: Implement htool_binary_parser()
This commit primarily implements the function htool_binary_parser()
for detecting whether a file is either ELF or Mach-O, and then loading
it into a htool_binary_t structure.

The ELF loading hasn't been implemented yet, only file detection.

However, Mach-O has. The htool_binary_t structure has a few fields,
such as `macho_list` and `fat_info`. These are set depending on the
file and flags.

If the file to be loaded is a single Mach-O, then it's just added
to the macho_list on it's own, however if the file is a FAT archive,
the fat_info field is populated and the list is filled with all the
contained mach_t structs.

Signed-off-by: Harry Moulton <me@h3adsh0tzz.com>