Loader: Implement htool_binary_select_arch().
This commit implements the htool_binary_select_arch() function
for selecting a macho_t from a given htool_binary_t structure and
arch name.

It will loop through the macho list in the htool_binary_t struct
until one matches the given arch name. If nothing matches, then
NULL is returned.

The arch-name MUST be exactly the same as the ones defined in the
macho loader.

Signed-off-by: Harry Moulton <me@h3adsh0tzz.com>