Mach-O: Implement support for the Thread State Load Command.
This commit implements support for the thread state load command. This
load command contains a flavour and count value that identify what
architecture and endian the thread state struct is, and then directly
following the thread_state command in memory is the cpu thread data.

It's stored as an array of uint64_t's, representing the general purpose
registers on AArch64, with system registers following.

Currently, only support for 64-bit thread commands is implemented, partly
due to me not being able to find a 32-bit decrypted kernel file to test.

Signed-off-by: Harry Moulton <me@h3adsh0tzz.com>