Mach-O: Symbols: Implement support for the --symbols option.
This commit implements support for iterating the static symbols table
within a Mach-O file under the LC_SYMTAB Load Command.

This functionality is primarily implemented in the nm.c file, where
the code is based on that of cctools nm.c.

A number of definitions have been added to commands/macho.h to support
debug symbols.

Signed-off-by: Harry Moulton <me@h3adsh0tz.com>