Misc: Use stdlib exit codes rather the htool_return_t
Because the htool_return_t type has success as '1', and failure as '0',
using these as exit codes means that the command line is assuming a good
run was a failure. This is an issue for CI too that fails once a command
exits with a value other than 0.

Anywhere the return value effects the exit code has been changed to use
the EXIT_FAILURE and EXIT_SUCCESS values from stdlib.h

Signed-off-by: Harry Moulton <me@h3adsh0tzz.com>