Analyse: Add initial functionality for detecting the Kernel Cache.
This commits adds a number of things, primarily the ability to detect a
kernel cache that has been passed to HTool. The HTool Loader has been
modified to now a) try to detect if a verified Mach-O is also a kernel
by looking for a __PRELINK_INFO segment and "Darwin Kernel Version"
string, and b) if the file is not a Mach-O or FAT, trying to detect if
it's an Image4. Eventually, iBoot/SEP/SecureROM detection will be done
after the Image4 check.

The few lines that load the client->filename have also been simplified,
as well as adding a TODO file so i can note anything that I want to add
in the future.

Signed-off-by: Harry Moulton <me@h3adsh0tzz.com>