Loader: Add pointer to firmware in htool_binary_t
This commit adds a `firmware` property to the htool_binary_t structure.

This works by setting bin->firmware to a void* pointer of whatever
structure it originally is. An example of this would be for xnu_t:

    * An xnu_t structure exists. To set this into the htool_binary_t,
      bin->firmware = (void *) xnu;
    * So bin->firmware can be identified, bin->flags =
      HTOOL_BINARY_FIRMWARETYPE_KERNEL.

The aim of this is to avoid having seperate fields for xnu, iboot, etc,
instead just having a pointer with the type identified by a flag.

Signed-off-by: Harry Moulton <me@h3adsh0tzz.com>