KEXT: Implement KEXT parsing for Split-style Kernels.
This commit implements support for the Split-style kernel KEXT format.
The Split-style format is used from iOS 9 to iOS 11. The Kernel Mach-O
contains an XML in the __PRELINK_INFO segment that maps out the location
of all the KEXTs, which are stored in the __PRELINK_TEXT segment.

However, the KEXTs are not stored together. Instead, the header plus the
load commands are stored in the __PRELINK_TEXT segment, which the
LC_SEGMENT_64 commands pointing to different areas of the Kernel Mach-O.

This creates a problem, as the KEXT Mach-O's will be to be rebuilt to
allow for them to be extracted and loaded into an application such as
IDA.

Signed-off-by: Harry Moulton <me@h3adsh0tzz.com>