Kernel: Implement support for all three types of kernel cache
This commit completes, with one exception, the support for the three
types of Kernel cache: Split, Merged and Fileset, for 64-bit Kernels.

The exception is that the KEXT Mach-O's are not complete. The offsets
point to areas within the Kernel Mach-O, not it's own. So if one were to
extract a KEXT, the __TEXT_EXEC, and other sections, along with data
such as symbols doesn't exist in the KEXT file.

The solution here is to implement functionality in Libhelper to repair
Mach-O files when they're extracted.

The last bit of support for KEXTs that needs to be implemented is
extracting KEXTs, either all of them or by Bundle Identifier.

Signed-off-by: Harry Moulton <me@h3adsh0tzz.com>