BUGFIX: Fix an issue with KEXTs keeping the same name (iOS 11).
This commit fixes an issue when dealing with the Kernel Extensions on
iOS 11 Kernels. The issue being that every KEXT was appearing with the
same name when running `analyse` with the `-l` flag.

The issue was caused by the `kext_name` value in
xnu_load_kext_list_split_style() being overwritten every time the loop
made a pass to look for the next KEXT. What this resulted in was all the
KEXTs in xnu->kexts being the same name (although different values for
address, etc) because they all had the same pointer.

This was addressed by making `kext_name` a pointer, rather than a char
array, calloc'ing and memset()'ing it at the start of every loop run,
and freeing at the end.

Finally, we strdup() the string when assigning it to the kext_t struct,
therefore there shouldn't be any kexts with the same name.

FIXES: HTOOL3-51

Signed-off-by: Harry Moulton <me@h3adsh0tzz.com>