Disassember: Initial implementation of Libarch in HTool
This commit adds the initial `disass` command and functionality for
disassembling most Mach-O files. Currently the disassembler works by
looking for the base offset of the __TEXT segment and creating libarch
instruction_t structs for every 8-bytes up to the end of __TEXT.

The output is also symbolicated, providing there is an LC_SYMTAB load
command that provides a symbol table. This is done by (mostly)
ustilising a hashmap, entering each symbols name and virtual address
into the map. When pritning each instruction, it checks if it's program
counter/addr value matches one of the symbols, and prints the symbol
name.

There is still a lot to do here, for example:

 * Only supports Mach-O's with executable code in __TEXT, meaning no
   support for the kernel.
 * No support for raw binaries.
 * Segment/section headers need to be printed as well as symbol info.
 * Handling of raw strings should be similar to Radare2.
 * Cannot specify base address to disassemble from.
 * ... and more.

Signed-off-by: Harry Moulton <me@h3adsh0tzz.com>