Initial Libarch integration (#10)
* Intiial libarch integration and setup

* BUGFIX: Fix an issue with KEXTs keeping the same name (iOS 11).

This commit fixes an issue when dealing with the Kernel Extensions on
iOS 11 Kernels. The issue being that every KEXT was appearing with the
same name when running `analyse` with the `-l` flag.

The issue was caused by the `kext_name` value in
xnu_load_kext_list_split_style() being overwritten every time the loop
made a pass to look for the next KEXT. What this resulted in was all the
KEXTs in xnu->kexts being the same name (although different values for
address, etc) because they all had the same pointer.

This was addressed by making `kext_name` a pointer, rather than a char
array, calloc'ing and memset()'ing it at the start of every loop run,
and freeing at the end.

Finally, we strdup() the string when assigning it to the kext_t struct,
therefore there shouldn't be any kexts with the same name.

FIXES: HTOOL3-51

Signed-off-by: Harry Moulton <me@h3adsh0tzz.com>

* Disassember: Initial implementation of Libarch in HTool

This commit adds the initial `disass` command and functionality for
disassembling most Mach-O files. Currently the disassembler works by
looking for the base offset of the __TEXT segment and creating libarch
instruction_t structs for every 8-bytes up to the end of __TEXT.

The output is also symbolicated, providing there is an LC_SYMTAB load
command that provides a symbol table. This is done by (mostly)
ustilising a hashmap, entering each symbols name and virtual address
into the map. When pritning each instruction, it checks if it's program
counter/addr value matches one of the symbols, and prints the symbol
name.

There is still a lot to do here, for example:

 * Only supports Mach-O's with executable code in __TEXT, meaning no
   support for the kernel.
 * No support for raw binaries.
 * Segment/section headers need to be printed as well as symbol info.
 * Handling of raw strings should be similar to Radare2.
 * Cannot specify base address to disassemble from.
 * ... and more.

Signed-off-by: Harry Moulton <me@h3adsh0tzz.com>

* Bump libhelper

* version bump

---------

Signed-off-by: Harry Moulton <me@h3adsh0tzz.com>
Co-authored-by: Harry Moulton <h3@Harrys-MacBook-Pro.local>
Co-authored-by: Harry Moulton <h3@Harrys-MBP.localdomain>