FIX: Resolve issue with --arch not selecting the correct architecture.
There has been an issue with the --arch flag where the provided string
was not selecting the correct architecture within a FAT file.

When checking if the --arch was given, htool_macho_select_arch() was
using htool_macho_check_fat(), which returns true if the file is a FAT
archive AND --arch isn't provided, therefore within
htool_macho_select_arch the condition kept failing and returning the
first element in the bin->machos list.

Now, the condition in htool_macho_select_arch() checks the client->opt
flags for --arch, and client->bin->flags for the filetype.

Addresses: Issue #12

Signed-off-by: Harry Moulton <me@h3adsh0tzz.com>